add Groovy Template view layer
added job.tpl and layout-main.tpl groovy templates.

Modified JobController.groovy to add the /job URI which directs to the
job.tpl view.

Modified build.gradle to add devtools and update the bootRun task.  With
these in place, when we run 'gradle bootRun' the application will start
and become available and any changes we make to static content will be
reloaded.  So we can change the templates and see the changes
immediately.